fixed up styles
