+++ src/App.vue
@@ -32,6 +32,8 @@ $frameWidth: $Frame-Width;
 $textColour: $Text-Colour;
 $backgroundColour: $Background-Colour;
 $fontFamily: $Font-Family;
+$highlightColour: $Highlight-Colour;
+$shadowColour: $Shadow-Colour;
 
 $footerPadding: 60px;
 
@@ -85,6 +87,65 @@ $footerPadding: 60px;
 		position: absolute;
 		top: 50%;
 		transform: translateY(-50%);
+	}
+
+}
+
+.AppPreview {
+	height: 350px;
+	width: 100%;
+	background: $highlightColour;
+
+	&__overlay {
+		width: 100%;
+		height: 100%;
+		background: $shadowColour;
+		opacity: 0.1;
+		transition: opacity 0.2s ease-in;
+
+		&:hover {
+			opacity: 0.9;
+		}
+	}
+
+	&__third {
+		height: 100%;
+		width: 33.33%;
+		background-size: cover;
+		background-position: center;
+		float: left;
+		position: relative;
+
+		&--published {
+			background-image: url('/static/images/published.jpg');
+		}
+
+		&--performed {
+			background-image: url('/static/images/performed.jpg');
+		}
+
+	}
+
+	&__text {
+		color: $highlightColour;
+	}
+
+	&__title {
+		position: absolute;
+		top: 50%;
+		left: 50%;
+		transform: translate(-50%, -50%);
+		font-size: 2rem;
+		font-weight: bold;
+		text-transform: uppercase;
+	}
+
+	&__caption {
+		width: 80%;
+		margin: 0 10%;
+		text-align: center;
+		position: absolute;
+		bottom: 10%;
 	}
 
 }
